+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # [develop](https://github.com/benlangfeld/ruby_speech)
+  * Allow wrapping a pre-parsed XML node nested arbitrary deeply as an NLSML document
 
 # [2.1.1](https://github.com/benlangfeld/ruby_speech/compare/v2.1.0...v2.1.1) - [2013-05-09](https://rubygems.org/gems/ruby_speech/versions/2.1.1)
   * Bugfix: Support numeric SISR literal tags